backend/requirements.txt
fastapi
uvicorn
scikit-learn
pandas
joblib
pydantic
shap
google-genai